moved PayloadSerializer to mqtt Package to resolve dependency not found
