+++ sensor-things-api-mqtt/src/main/java/org/n52/sta/mqtt/handler/PayloadSerializer.java
@@ -31,14 +31,11 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.n52.sta.service.serializer;
+package org.n52.sta.mqtt.handler;
 
 import com.google.common.io.ByteStreams;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Set;
 import org.apache.olingo.commons.api.data.ContextURL;
 import org.apache.olingo.commons.api.data.Entity;
 import org.apache.olingo.commons.api.edm.EdmEntitySet;
@@ -50,9 +47,14 @@ import org.apache.olingo.server.api.serializer.SerializerException;
 import org.apache.olingo.server.api.serializer.SerializerResult;
 import org.n52.sta.service.query.QueryOptions;
 import org.n52.sta.service.query.QueryOptionsHandler;
+import org.n52.sta.service.serializer.SensorThingsSerializer;
 import org.n52.sta.utils.EntityAnnotator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Set;
 
 /**
  * Payload serializer for Entities
